Refactor DistributedCircuit
